refactor: replace OpenAL audio with javax.sound FloatControl and fix pause/resume sync
  - Switch volume control from OpenAL to FloatControl.MASTER_GAIN (dB conversion from 0–100 linear scale)
  - Fix audio/video desync on resume by tracking pauseStart timestamp to offset startTime
  - Wire playPause and volume action commands to BackgroundPanel (were empty stubs)
  - Add VolumeSliderListener for dedicated slider change handling
  - Add popup debounce (300ms) to prevent immediate reopening on volumeBtn click
  - Change volume type int → double for precision
  - Rename ButtonHandler → ButtonsListener, MainWindowsHandler → WindowsListener

+++ src/main/java/fr/guillaumewlt/parser/SettingsJSONParser.java
@@ -18,7 +18,7 @@ public class SettingsJSONParser {
     private String maxRam;
     private String version;
     private String username;
-    private int volume = 30;
+    private double volume = 30;
     private boolean videoPaused = false;
 
     public SettingsJSONParser(LauncherContext context) {
@@ -35,7 +35,7 @@ public class SettingsJSONParser {
                 maxRam = jsonObj.optString("maxRam", "2048");
                 version = jsonObj.optString("version", null);
                 username = jsonObj.optString("username", "Player");
-                volume = jsonObj.optInt("volume", 30);
+                volume = jsonObj.optDouble("volume", 30);
                 videoPaused = jsonObj.optBoolean("videoPaused", false);
             } catch (IOException ex) {
                 ex.printStackTrace();
